Moves kubectl type to its own file under kubeutil.
